gomod: update x/tools for go1.18 support
Before this commit go-mockgen fails on go1.18 code with errors like:

  internal error: package "context" without types was imported from "package/name/here"

This commit updates the dependency. After this update go-mockgen works.

Note: the dependencies in this repository are important. In newer
versions of go if you run "go install go-mockgen@version" it uses the
module file in this repository, not your repository.

Note: we update all dependencies. We can limit this just to the relevant
deps.